Quick debug before starting to create the Client

+++ book-service/src/main/java/com/openclassroom/bookservice/Controller/BookController.java
@@ -11,6 +11,8 @@ import com.openclassroom.bookservice.Service.BookService;
 import com.openclassroom.bookservice.Service.UserService;
 import com.openclassroom.bookservice.Service.LoanService;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -23,7 +25,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 @RestController
 @RequestMapping(value = "library/")
 public class BookController {
-
+    //private static final Logger logger = LoggerFactory.getLogger(BookController.class);
     @Autowired
     private BookService bookService;
 
@@ -75,8 +77,10 @@ public class BookController {
             return ResponseEntity.noContent().build();
         }
 
-        if(bookService.findByTitle(book.getTitle()) == null)
+        if(bookService.findByTitle(book.getTitle()).isEmpty()) {
             bookService.save(book);
+        }
+            
 
         URI location = ServletUriComponentsBuilder
                         .fromCurrentContextPath()
@@ -89,7 +93,7 @@ public class BookController {
     // TO-DO add a borrower only if the book is in stock (IF IT HIS the microservice's job) (should be the web app)
     @RequestMapping(value = "book/{bookId}/borrower/add", method = RequestMethod.POST)
     public ResponseEntity<Loan> addBorrower(@PathVariable(name = "bookId") String bookId, @RequestBody User user) {
-        
+        Loan loan;
         if (user == null) {
             return ResponseEntity.noContent().build();
         }
@@ -97,21 +101,28 @@ public class BookController {
         if (userService.findByName(user.getName()) == null)
             userService.save(user);
         
-        Loan loan = new Loan(bookService.findById(bookId).get(), user);
-        loanService.save(loan);
+        if (bookService.findById(bookId).isPresent()) {
+            loan = new Loan(bookService.findById(bookId).get(), user);
+            loanService.save(loan);
 
-        URI location = ServletUriComponentsBuilder
-                            .fromCurrentContextPath()
-                            .path("/{id}")
-                            .buildAndExpand(loan.getId())
-                            .toUri();
-        return ResponseEntity.created(location).build();
+            URI location = ServletUriComponentsBuilder
+                                .fromCurrentContextPath()
+                                .path("/{id}")
+                                .buildAndExpand(loan.getId())
+                                .toUri();
+            return ResponseEntity.created(location).build();
+        }
+
+        else {
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     // Working fine (done by strict User name matching) and can delete only if one of the same entity exists.
     @RequestMapping(value = "book/{bookId}/borrower/delete", method = RequestMethod.POST)
-    public void deleteBorrower(@PathVariable(name="bookId") String bookId, @RequestBody User user){
-        loanService.delete(loanService.findByUser(user));
+    public void deleteBorrower(@RequestBody User user){
+        if (loanService.findByUser(user) != null)
+            loanService.delete(loanService.findByUser(user));
     }
 
     //TO-DO
